+++ SuperService/Objects/EventServicesMaterialsLine.cs
@@ -20,11 +20,11 @@ namespace Test
         public string Ref { get; set; }
         public string SKU { get; set; }
         // TODO: покурить с даблом, что бы не нарваться на проблему округления
-        public double Price { get; set; }
-        public double AmountPlan { get;  set; }
-        public double SumPlan { get; set; }
-        public double AmountFact { get; set; }
-        public double SumFact { get; set; }
+        public decimal Price { get; set; }
+        public decimal AmountPlan { get;  set; }
+        public decimal SumPlan { get; set; }
+        public decimal AmountFact { get; set; }
+        public decimal SumFact { get; set; }
 
     }
 }
